Refactor segregated intersection classification to right module
