+++ include/extractor/extractor.hpp
@@ -47,7 +47,6 @@ namespace extractor
 
 class ScriptingEnvironment;
 struct ProfileProperties;
-class NodeBasedGraphFactory;
 
 class Extractor
 {
@@ -102,13 +101,6 @@ class Extractor
     void WriteConditionalRestrictions(
         const std::string &path,
         std::vector<ConditionalTurnRestriction> &conditional_turn_restrictions);
-
-    // Find all "segregated" edges, e.g. edges that can be skipped in turn instructions.
-    // The main cases are:
-    // - middle edges between two osm ways in one logic road (U-turn)
-    // - staggered intersections (X-cross)
-    // - square/circle intersections
-    std::unordered_set<EdgeID> FindSegregatedNodes(NodeBasedGraphFactory &factory);
 };
 }
 }
